refactor(src): 🔨 Move ads outside of file
Move ads outside of file.
